refactor(search): Moved views in self package
