+++ REES46/Classes/Stories/StoryViewController/SearchWidget/Presentation/SearchWidgetCategoriesButton.swift
