+++ REES46/Classes/Stories/StoryViewController/SearchWidget/Presentation/SearchWidgetDelegate.swift
