+++ REES46/Classes/Stories/StoryViewController/SearchWidget/Presentation/SearchWidgetHistoryButton.swift
