+++ REES46/Classes/Stories/StoryViewController/SearchWidget/Presentation/SearchWidgetModel.swift
