+++ REES46/Classes/Stories/StoryViewController/SearchWidget/Protocol/SearchResultsViewDelegate.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+public protocol SearchResultsViewDelegate: AnyObject {
+     func didSelectResult(_ result: SearchResult)
+}
